iot-predictive-maintenance/requirements.txt
Flask
kafka-python
pandas
scikit-learn
joblib
boto3
pyspark

